+++ src/lib/pages/pools/components/pool-details/tables/pool-assets/PoolAssetsTableHeader.tsx
@@ -2,12 +2,12 @@ import type { GridProps } from "@chakra-ui/react";
 import { Grid } from "@chakra-ui/react";
 
 import { TableHeader } from "lib/components/table";
-import type { PoolDetail } from "lib/types";
+import type { PoolData } from "lib/types";
 import { PoolType } from "lib/types";
 
 interface PoolHeaderProps {
-  poolType: PoolDetail["type"];
-  isSupported: PoolDetail["isSupported"];
+  poolType: PoolData["type"];
+  isSupported: PoolData["isSupported"];
   templateColumns: GridProps["templateColumns"];
 }