+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/messages/components/PoolLogoLink.tsx
@@ -2,21 +2,16 @@ import { Flex } from "@chakra-ui/react";
 
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { PoolLogo } from "lib/pages/pools/components/PoolLogo";
-import type { PoolDetail } from "lib/types";
+import type { PoolData } from "lib/types";
 
 interface PoolLogoLinkProps {
-  pool: PoolDetail;
+  pool: PoolData;
   ampCopierSection?: string;
 }
 
 export const PoolLogoLink = ({ pool, ampCopierSection }: PoolLogoLinkProps) => (
   <Flex gap={1}>
-    <PoolLogo
-      tokens={pool.poolLiquidity}
-      logoSize={5}
-      marginLeft={-4}
-      minW={0}
-    />
+    <PoolLogo tokens={pool.liquidity} logoSize={5} marginLeft={-4} minW={0} />
     <ExplorerLink
       type="pool_id"
       value={pool.id.toString()}