+++ src/lib/pages/pools/poolId.tsx
@@ -6,7 +6,7 @@ import { usePoolConfig, useTierConfig } from "lib/app-provider";
 import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { CelatoneSeo } from "lib/components/Seo";
-import { InvalidState } from "lib/components/state";
+import { ErrorFetching, InvalidState } from "lib/components/state";
 import { UserDocsLink } from "lib/components/UserDocsLink";
 
 import {
@@ -14,21 +14,22 @@ import {
   PoolRelatedTxs,
   PoolTopSection,
 } from "./components/pool-details";
-import { usePool } from "./data";
+import { useDerivedPoolData } from "./data";
 import { zPoolDetailsQueryParams } from "./types";
 
 const InvalidPool = () => <InvalidState title="Pool does not exist" />;
 
 const PoolIdBody = ({ poolId }: { poolId: number }) => {
   const router = useRouter();
-  const { pool, isLoading } = usePool(poolId);
+  const { pool, isLoading } = useDerivedPoolData(poolId);
 
   useEffect(() => {
     if (router.isReady) track(AmpEvent.TO_POOL_DETAILS);
   }, [router.isReady]);
 
   if (isLoading) return <Loading />;
-  if (!pool) return <InvalidPool />;
+  if (pool === undefined) return <ErrorFetching dataName="pool information" />;
+  if (pool === null) return <InvalidPool />;
 
   return (
     <>